backend/requirements.txt
# Minimale Requirements für Push Server (ohne Docker)
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.3.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Optional für Admin UI
streamlit==1.28.2  # Nur wenn du die Python Admin UI willst

# Für Produktion später hinzufügen:
# sqlalchemy==2.0.23
# redis==5.0.1
# psycopg2-binary==2.9.9